+++ src/components/Header/OptionalHeader.tsx
@@ -1,6 +1,6 @@
 import { getCookie } from "src/utils/CookieSetting";
-import AuthNav from "./GNB/AuthNav";
-import GuestNav from "./GNB/GuestNav";
+import AuthNav from "./GNB/AuthNav/AuthNav";
+import GuestNav from "./GNB/GuestNav/GuestNav";
 
 const OptionalHeader = () => {
   const accessToken = getCookie("accessToken");
